Fixing crash on Context destruct with a hack, in-place new for pool
When objects in the pool are destroyed and returned to be claimed by a succeeding create call, they are explicitly destructed and re-constructed at the same memory address. Hence all cleanup code for pooled objects can live in the destructor, and there is no need for a special cleanup function.