Full multi-texturing support
Re-arranged the Sprite class to support multi-frame and multi-textured sprites (e.g., each animation frame can have multiple textures for each type of texture (such as Diffuse, Specular, etc.)). The Sprite class uses a mix value per texture index to tell the shader how much to mix when multi-texturing. When there are no more textures to mix, the Sprite class returns a mix value of 1.0, retaining the prior mix data so as the shader continues to sample the rest of the texture units, that data is ignored.

Former-commit-id: fcc1012eb87c3484d1ff7b2a353665c5571416af